Added class for advanced operations (exponents and square roots)

+++ basic.cpp
@@ -3,32 +3,32 @@
 using namespace std;
 
 // add() function implementation - all function declarations located in basic.h
-int basic::add(int num1, int num2) {
-    int num3;
+int basic::add(float num1, float num2) {
+    float num3;
     num3 = num1 + num2;
     cout << num3;
     return 0;
 }
 
 // subtract() function implementation
-int basic::subtract(int num1, int num2) {
-    int num3;
+int basic::subtract(float num1, float num2) {
+    float num3;
     num3 = num1 - num2;
     cout << num3;
     return 0;
 }
 
 // multiply() function implementation
-int basic::multiply(int num1, int num2) {
-    int num3;
+int basic::multiply(float num1, float num2) {
+    float num3;
     num3 = num1 * num2;
     cout << num3;
     return 0;
 }
 
 // divide() function implementation
-int basic::divide(int num1, int num2) {
-    int num3;
+int basic::divide(float num1, float num2) {
+    float num3;
     num3 = num1 / num2;
     cout << num3;
     return 0;
